testing

+++ lib/utils.h
@@ -47,6 +47,6 @@ void printHashTables(int dim, int l, int m, double ***tables);
 
 int printHashBuckets(int dim, int l, int m, HashBucket *buckets);
 
-int printHashValues(int l, int m, double **hashValue);
+int printHashValues(int l, int m, int **hashValue);
 
 #endif //LSH_PROBING_UTILS_H
